+++ src/main/java/by/gomselmash/aspiski/repository/MachineRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MachineRepository extends JpaRepository<Machine, Integer> {
+public interface MachineRepository extends JpaRepository<Machine, Long> {
     List<Machine> findAllByOrderByNameAsc();
     boolean existsByNameIgnoreCase(String name);
-    boolean existsByMachineType_Id(int machineTypeId);
+    boolean existsByMachineType_Id(Long machineTypeId);
 }
