+++ src/main/java/by/gomselmash/aspiski/repository/MachineTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MachineTypeRepository extends JpaRepository<MachineType, Integer> {
+public interface MachineTypeRepository extends JpaRepository<MachineType, Long> {
     List<MachineType> findAllByOrderByNameAsc();
     boolean existsByNameIgnoreCase(String name);
 }
