+++ src/main/java/by/gomselmash/aspiski/repository/ProgramRepository.java
@@ -8,12 +8,12 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface ProgramRepository extends JpaRepository<Program, Integer> {
+public interface ProgramRepository extends JpaRepository<Program, Long> {
     List<Program> findAllByOrderByPartNumberAsc();
     List<Program> findAllByDateIsBetweenOrderByDateAscDeveloperAscMachine_MachineType_NameAsc(LocalDate from, LocalDate to);
     boolean existsByProgramNumberIgnoreCase(String programNumber);
-    boolean existsByWorkshop_Id(int workshopId);
-    boolean existsByDeveloper_Id(int developerId);
-    boolean existsByMachine_Id(int machineId);
-    boolean existsByControlSystem_Id(int controlSystemId);
+    boolean existsByWorkshop_Id(Long workshopId);
+    boolean existsByDeveloper_Id(Long developerId);
+    boolean existsByMachine_Id(Long machineId);
+    boolean existsByControlSystem_Id(Long controlSystemId);
 }
